# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: R&S Admissao API
  description: |
    API do sistema de Recrutamento e Selecao (R&S) para agendamento de aulas teste.

    Este sistema permite que candidatos agendem aulas teste em unidades educacionais,
    enquanto a equipe de recrutamento gerencia vagas, configuracoes de horarios e
    bloqueios de agenda.

    ## Autenticacao

    A API utiliza dois mecanismos de autenticacao:

    - **JWT Bearer Token**: Para usuarios autenticados (equipe de recrutamento)
    - **Acesso Publico**: Alguns endpoints sao publicos para candidatos

    Endpoints marcados com `optionalAuth` podem ser acessados com ou sem autenticacao,
    com comportamento diferenciado baseado no contexto.

    ## Ambientes

    | Ambiente | URL Base |
    |----------|----------|
    | Producao | https://api.rsadmissao.tomeducacao.com.br/api/v1 |
    | Homologacao | https://api-homolog.rsadmissao.tomeducacao.com.br/api/v1 |

    ## Versionamento

    A API segue versionamento semantico. A versao atual e `v1`.

    ## Rate Limiting

    A API implementa limitacao de taxa para prevenir abuso:

    | Endpoint | Limite | Janela |
    |----------|--------|--------|
    | `POST /auth/login` | 5 requisicoes | 1 minuto |
    | `POST /booking` (publico) | 10 requisicoes | 1 minuto |
    | `GET /availability` | 60 requisicoes | 1 minuto |
    | Demais endpoints | 100 requisicoes | 1 minuto |

    **Headers de resposta:**
    - `X-RateLimit-Limit`: Limite total de requisicoes
...

...

...

...

...
servers:
...
paths:
...
  /schedule:
    get:
      tags:
        - schedule
      summary: Consultar configuracao de horarios
      description: |
        Retorna a configuracao de horarios para uma unidade ou a configuracao global.

        **Hierarquia de configuracao:**
        1. Se `id_unidade` for informado e existir config especifica, retorna ela
        2. Se `id_unidade` for informado mas nao houver config especifica, retorna global
        3. Se `id_unidade` nao for informado, retorna global

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: getScheduleConfig
      security:
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: false
          description: |
            ID da unidade para buscar configuracao especifica.
            Se omitido, retorna a configuracao global.
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Configuracao retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              examples:
                global:
                  summary: Configuracao global
                  value:
                    success: true
                    config:
                      id_schedule_config: 1
                      id_unidade: null
                      morning_start_at: "08:00:00"
                      morning_end_at: "12:00:00"
                      afternoon_start_at: "13:00:00"
                      afternoon_end_at: "17:00:00"
                      slot_size: 50
                      d_rule_start: 1
                      d_rule_end: 30
                      valid_from: null
                      valid_until: null
                      active: true
                    is_global: true
                unidade:
                  summary: Configuracao especifica de unidade
                  value:
                    success: true
                    config:
                      id_schedule_config: 5
                      id_unidade: 123
                      morning_start_at: "07:30:00"
                      morning_end_at: "11:30:00"
                      afternoon_start_at: "13:30:00"
                      afternoon_end_at: "17:30:00"
                      slot_size: 45
                      d_rule_start: 2
                      d_rule_end: 15
                      valid_from: "2025-01-01"
                      valid_until: "2025-06-30"
                      active: true
                    is_global: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Configuracao nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Configuracao de horarios nao encontrada"
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - schedule
      summary: Criar ou atualizar configuracao de horarios
      description: |
        Cria ou atualiza a configuracao de horarios.

        **Comportamento:**
        - Se `id_unidade` for `null` ou omitido, afeta a configuracao global
        - Se `id_unidade` for informado, cria/atualiza config especifica da unidade
        - Operacao de upsert: cria se nao existir, atualiza se existir

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: upsertScheduleConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleConfigRequest'
            examples:
              global:
                summary: Configuracao global
                value:
                  id_unidade: null
                  morning_start_at: "08:00:00"
                  morning_end_at: "12:00:00"
                  afternoon_start_at: "13:00:00"
                  afternoon_end_at: "17:00:00"
                  slot_size: 50
                  d_rule_start: 1
                  d_rule_end: 30
              unidade:
                summary: Configuracao especifica de unidade
                value:
                  id_unidade: 123
                  morning_start_at: "07:30:00"
                  morning_end_at: "11:30:00"
                  afternoon_start_at: "13:30:00"
                  afternoon_end_at: "17:30:00"
                  slot_size: 45
                  d_rule_start: 2
                  d_rule_end: 15
                  valid_from: "2025-01-01"
                  valid_until: "2025-06-30"
      responses:
        '200':
          description: Configuracao atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              example:
                success: true
                message: "Configuracao atualizada com sucesso"
                config:
                  id_schedule_config: 5
                  id_unidade: 123
                  morning_start_at: "07:30:00"
                  morning_end_at: "11:30:00"
                  afternoon_start_at: "13:30:00"
                  afternoon_end_at: "17:30:00"
                  slot_size: 45
                  d_rule_start: 2
                  d_rule_end: 15
                  valid_from: "2025-01-01"
                  valid_until: "2025-06-30"
                  active: true
        '201':
          description: Configuracao criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              example:
                success: true
                message: "Configuracao criada com sucesso"
                config:
                  id_schedule_config: 6
                  id_unidade: 456
                  morning_start_at: "08:00:00"
                  morning_end_at: "12:00:00"
                  afternoon_start_at: "13:00:00"
                  afternoon_end_at: "17:00:00"
                  slot_size: 50
                  d_rule_start: 1
                  d_rule_end: 30
                  valid_from: null
                  valid_until: null
                  active: true
        '400':
          description: Dados invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errors:
                  - msg: "slot_size deve estar entre 15 e 120 minutos"
                    param: "slot_size"
                    location: "body"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /schedule/{id_unidade}:
    delete:
      tags:
        - schedule
      summary: Remover configuracao de unidade
      description: |
        Remove a configuracao especifica de uma unidade.

        Apos a remocao, a unidade passa a usar a configuracao global.

        **Nota:** Nao e possivel remover a configuracao global (id_unidade = null).

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: deleteScheduleConfig
      security:
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: path
          required: true
          description: ID da unidade cuja configuracao sera removida
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Configuracao removida com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Configuracao removida. A unidade agora usa a configuracao global."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Configuracao nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Configuracao especifica nao encontrada para esta unidade"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # SCHEDULE-BLOCK
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido atraves do endpoint `/auth/login`.

        Enviar no header: `Authorization: Bearer {token}`
  responses:
    Unauthorized:
      description: Token nao fornecido ou invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Token nao fornecido ou invalido"
    Forbidden:
      description: Usuario nao tem permissao para esta acao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Acesso negado. Permissao insuficiente."
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Erro interno do servidor"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Mensagem de erro
          example: "Descricao do erro"
        errors:
          type: array
          description: Lista de erros de validacao
          items:
            type: object
            properties:
              msg:
                type: string
                example: "Campo obrigatorio"
              param:
                type: string
                example: "id_unidade"
              location:
                type: string
                example: "query"
        details:
          type: string
          description: Detalhes adicionais do erro (desenvolvimento)

    # --------------------------------------------------------------------------
    # Availability Schemas
    # --------------------------------------------------------------------------
...
    ScheduleConfig:
      type: object
      required:
        - id_schedule_config
        - morning_start_at
        - morning_end_at
        - afternoon_start_at
        - afternoon_end_at
        - slot_size
        - d_rule_start
        - d_rule_end
        - active
      properties:
        id_schedule_config:
          type: integer
          description: ID da configuracao
          example: 1
        id_unidade:
          type: integer
          nullable: true
          description: ID da unidade (null = configuracao global)
          example: 123
        morning_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da manha
          example: "08:00:00"
        morning_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da manha
          example: "12:00:00"
        afternoon_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da tarde
          example: "13:00:00"
        afternoon_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da tarde
          example: "17:00:00"
        slot_size:
          type: integer
          minimum: 15
          maximum: 120
          description: Duracao do slot em minutos
          example: 50
        d_rule_start:
          type: integer
          minimum: 0
          description: Dias minimos de antecedencia para agendamento
          example: 1
        d_rule_end:
          type: integer
          minimum: 1
          description: Dias maximos de antecedencia para agendamento
          example: 30
        valid_from:
          type: string
          format: date
          nullable: true
          description: Data de inicio da validade da configuracao
          example: "2025-01-01"
        valid_until:
          type: string
          format: date
          nullable: true
          description: Data de fim da validade da configuracao
          example: "2025-03-31"
        active:
          type: boolean
          description: Indica se a configuracao esta ativa
          example: true
    ScheduleConfigRequest:
      type: object
      required:
        - morning_start_at
        - morning_end_at
        - afternoon_start_at
        - afternoon_end_at
        - slot_size
        - d_rule_start
        - d_rule_end
      properties:
        id_unidade:
          type: integer
          nullable: true
          description: ID da unidade (null = configuracao global)
          example: 123
        morning_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da manha
          example: "08:00:00"
        morning_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da manha
          example: "12:00:00"
        afternoon_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da tarde
          example: "13:00:00"
        afternoon_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da tarde
          example: "17:00:00"
        slot_size:
          type: integer
          minimum: 15
          maximum: 120
          description: Duracao do slot em minutos
          example: 50
        d_rule_start:
          type: integer
          minimum: 0
          description: Dias minimos de antecedencia para agendamento
          example: 1
        d_rule_end:
          type: integer
          minimum: 1
          description: Dias maximos de antecedencia para agendamento
          example: 30
        valid_from:
          type: string
          format: date
          nullable: true
          description: Data de inicio da validade da configuracao
          example: "2025-01-01"
        valid_until:
          type: string
          format: date
          nullable: true
          description: Data de fim da validade da configuracao
          example: "2025-03-31"
        active:
          type: boolean
          default: true
          description: Indica se a configuracao esta ativa
          example: true
    ScheduleConfigResponse:
      type: object
      required:
        - success
        - config
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Configuracao atualizada com sucesso"
        config:
          $ref: '#/components/schemas/ScheduleConfig'
        is_global:
          type: boolean
          description: Indica se a configuracao retornada e global
          example: false

    # --------------------------------------------------------------------------
    # Schedule Block Schemas
    # --------------------------------------------------------------------------
...